#69 [STYLE] 친구 목록 페이지 수정
