+++ src/pages/FriendPage/FriendPage.jsx
@@ -1,9 +1,6 @@
-import { useQuery } from '@tanstack/react-query';
-
-import { getFriendList } from '../../apis/friend.js';
+import { Link, Outlet } from 'react-router-dom';
 
 import { AdComponent } from '../../components/AdComponent';
-import { Avatar } from '../../components/Avatar';
 import { Button } from '../../components/Button';
 import { RandomFriend } from '../../components/RandomFriend';
 
@@ -11,48 +8,29 @@ import AdImg from '../../dummy/images/cat1.png';
 
 import * as S from './FriendPage.style.jsx';
 
-const nickname = '댕댕산책가';
-
 export default function FriendPage() {
-  const { data } = useQuery({
-    queryKey: ['friendList'],
-    queryFn: () => getFriendList(),
-    staleTime: 1000 * 60 * 5,
-  });
-
-  if (!data) return null;
-
   return (
     <S.FriendLayout>
-      <S.FriendContainer>
-        <S.Title>{nickname}님의 친구 목록</S.Title>
-        <S.FriendList>
-          {data.map(({ id, nickname, profile_url: profileUrl }) => (
-            <S.FriendItem
-              key={id}
-              onClick={() => {
-                console.log(`${nickname}님 앨범 페이지로 이동!`);
-              }}
-            >
-              <Avatar src={profileUrl} size='66px' />
-              <span>{nickname}</span>
-              <Button
-                onClick={(event) => {
-                  event.stopPropagation();
-                  console.log(`${nickname} 친구 삭제`);
-                }}
-                width='100px'
-                padding='8px'
-                borderRadius='5px'
-                buttonStyle='white'
-                hasBorder
-              >
-                친구 삭제
-              </Button>
-            </S.FriendItem>
-          ))}
-        </S.FriendList>
-      </S.FriendContainer>
+      <S.MenuBox>
+        <Link to='/friend/list'>
+          <Button buttonStyle='white' hasBorder padding='13px'>
+            친구 목록
+          </Button>
+        </Link>
+        <Link to='/friend/sentList'>
+          <Button buttonStyle='white' hasBorder padding='13px'>
+            보낸 요청
+          </Button>
+        </Link>
+        <Link to='/friend/receivedList'>
+          <Button buttonStyle='white' hasBorder padding='13px'>
+            받은 요청
+          </Button>
+        </Link>
+      </S.MenuBox>
+      <S.Content>
+        <Outlet />
+      </S.Content>
       <S.RightSide>
         <RandomFriend />
         <AdComponent adImage={AdImg} adText='광고 컴포넌트입니다' />
