+++ src/pages/FriendPage/FriendReceivedList/FriendReceivedList.jsx
@@ -1,0 +1,3 @@
+export default function FriendReceivedList() {
+  return <div></div>;
+}
