+++ src/pages/FriendPage/FriendSentList/FriendSentList.jsx
@@ -1,0 +1,3 @@
+export default function FriendSentList() {
+  return <div></div>;
+}
