+++ src/pages/FriendPage/index.js
@@ -1,1 +1,4 @@
 export { default as FriendPage } from './FriendPage';
+export { default as FriendList } from './FriendList/FriendList';
+export { default as FriendSentList } from './FriendSentList/FriendSentList';
+export { default as FriendReceivedList } from './FriendReceivedList/FriendReceivedList';
